pyproject.toml
[tool.poetry]
name = "g1asm"
version = "0.1.4"
description = "Assembler implementation for the g1 ISA"
authors = ["7Limes"]
readme = "README.md"
license = "MIT"
repository = "https://github.com/7Limes/g1asm"
keywords = ["assembler", "cli"]


[tool.poetry.dependencies]
python = ">=3.10"
construct = ">=2.10.70"
pillow = ">=11.3.0"
rply = ">=0.7.8"


[tool.poetry.scripts]
g1a = "g1asm.assembler:main"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"